+++ commands.js
@@ -4,7 +4,9 @@ const taskEditHandler = require("./commands/taskEdit");
 const taskCompleteHandler = require("./commands/taskComplete");
 const taskDeleteHandler = require("./commands/taskDelete");
 const tasksHandler = require("./commands/tasks");
+
 const reminderActionsHandler = require("./commands/reminderActions");
+const listAdminsHandler = require("./commands/listadmins");
 
 module.exports = function (app, db) {
   taskHandler(app, db);
@@ -13,4 +15,5 @@ module.exports = function (app, db) {
   taskDeleteHandler(app, db);
   tasksHandler(app, db);
   reminderActionsHandler(app, db);
+  listAdminsHandler(app, db);
 };